ref: move some functions to helpers mod
